// src/modules/iE/reports/operation-bulletin-report/operation-bulletin-report-type.ts
export type OperationBulletinReportType = {
  DETAILS_ID: number;
  ID: number;
  SECTIONID: number;
  STYLEID: number;
  COMPANYID: number;
  ALLOWANCERATE: number;
  TARGERPERHOUR: number;
  TOTALSMV: number;
  TOTALREQUIREMP: number;
  TOTALALLOTTEDMP: number;
  EFFICIENCY: number;
  CREATEBY: string;
  STYLEIMAGE?: Uint8Array;
  STYLENO: string;
  STYLENAME: string;
  ITEMTYPE: string;
  BUYERNAME: string;
  SECTIONNAME: string;
  OPERATIONID: number;
  MACHINEID: number;
  SPINO: string;
  SEAMLENGTH: number;
  NOOFFREQUENCY: number;
  STANDARDCYCLETIME: number;
  RATING: number;
  SMV: number;
  CAPACITYHR: number;
  REQMP: number;
  ALLOTTEDMP: number;
  PLANWS: number;
  PRESSUREFOOT: string;
  MACHINENAME: string;
  GUIDEFOLDER: string;
  ATTACHMENT: string;
  REMARKS: string;
  FACCOMPANYNAME: string;
  FACCOMPANYADDRESS: string;
  OPERATIONNAME: string;
  MACHINECODE: string;
  COMPANYNAME: string;
  COMPANYADDRESS: string;
  COMPANYUTILITY: string;
  OPERATION_TYPE_ID: number;
  OPERATIONTYPE: string;
  MAINSECTION: string;
};